Made the architecture leader node failure resilient
I modified and created several methods, properties, etc, in other to make ratio consensus resilient to a failure of the leader node. Using min and max consensus, the leader and deputy are voted for during each ratio consensus, respectively. If a leader is unavailable after a schedule timeout is past, the deputy becomes the new leader